Address android-reviewer findings
- Use Properties.Resources.XA#### for error messages (XA4212, XA4213, XA4217)
  - Add Properties/Resources.resx + Resources.Designer.cs to TrimmableTypeMap project
  - Remove XA4212/XA4217 from Xamarin.Android.Build.Tasks resources (they are
    only used in TrimmableTypeMapGenerator, not in any Build.Tasks code path)

- Fix cross-platform path separator bugs in Trimmable.targets
  - Replace \**\*.java glob with /**/*.java
  - Replace android\src\ destination paths with android/src/
  - Replace ..\ TrimmerRootDescriptor path with ../

- Split 13 new types out of JavaPeerInfo.cs (one type per file)
  - ComponentInfo.cs  (ComponentKind + ComponentInfo)
  - IntentFilterInfo.cs
  - MetaDataInfo.cs
  - AssemblyManifestInfo.cs
  - ManifestAttributeInfo.cs  (Permission*, UsesPermission, UsesFeature,
                                UsesLibrary, UsesConfiguration, PropertyInfo)
  - TrimmableTypeMapTypes.cs  (ManifestConfig + TrimmableTypeMapResult)

- Replace string.IsNullOrEmpty() static calls with .IsNullOrEmpty() extension
  - Add NullableExtensions.cs to TrimmableTypeMap project
  - ManifestGenerator.cs: 6 occurrences
  - TrimmableTypeMapGenerator.cs: 3 occurrences

- Change List<string> to IReadOnlyList<string> in TrimmableTypeMapResult
