+++ src/Microsoft.Android.Sdk.TrimmableTypeMap/Scanner/JavaPeerInfo.cs
@@ -316,177 +316,3 @@ public enum ActivationCtorStyle
 	JavaInterop,
 }
 
-/// <summary>
-/// The kind of Android component (Activity, Service, etc.).
-/// </summary>
-public enum ComponentKind
-{
-	Activity,
-	Service,
-	BroadcastReceiver,
-	ContentProvider,
-	Application,
-	Instrumentation,
-}
-
-/// <summary>
-/// Describes an Android component attribute ([Activity], [Service], etc.) on a Java peer type.
-/// All named property values from the attribute are stored in <see cref="Properties"/>.
-/// </summary>
-public sealed record ComponentInfo
-{
-	/// <summary>
-	/// The kind of component.
-	/// </summary>
-	public required ComponentKind Kind { get; init; }
-
-	/// <summary>
-	/// All named property values from the component attribute.
-	/// Keys are property names (e.g., "Label", "Exported", "MainLauncher").
-	/// Values are the raw decoded values (string, bool, int for enums, etc.).
-	/// </summary>
-	public IReadOnlyDictionary<string, object?> Properties { get; init; } = new Dictionary<string, object?> ();
-
-	/// <summary>
-	/// Intent filters declared on this component via [IntentFilter] attributes.
-	/// </summary>
-	public IReadOnlyList<IntentFilterInfo> IntentFilters { get; init; } = [];
-
-	/// <summary>
-	/// Metadata entries declared on this component via [MetaData] attributes.
-	/// </summary>
-	public IReadOnlyList<MetaDataInfo> MetaData { get; init; } = [];
-
-	/// <summary>
-	/// Whether the component type has a public parameterless constructor.
-	/// Required for manifest inclusion — XA4213 error if missing.
-	/// </summary>
-	public bool HasPublicDefaultConstructor { get; init; }
-}
-
-/// <summary>
-/// Describes an [IntentFilter] attribute on a component type.
-/// </summary>
-public sealed record IntentFilterInfo
-{
-	/// <summary>
-	/// Action names from the first constructor argument (string[]).
-	/// </summary>
-	public IReadOnlyList<string> Actions { get; init; } = [];
-
-	/// <summary>
-	/// Category names.
-	/// </summary>
-	public IReadOnlyList<string> Categories { get; init; } = [];
-
-	/// <summary>
-	/// Named properties (DataScheme, DataHost, DataPath, Label, Icon, Priority, etc.).
-	/// </summary>
-	public IReadOnlyDictionary<string, object?> Properties { get; init; } = new Dictionary<string, object?> ();
-}
-
-/// <summary>
-/// Describes a [MetaData] attribute on a component type.
-/// </summary>
-public sealed record MetaDataInfo
-{
-	/// <summary>
-	/// The metadata name (first constructor argument).
-	/// </summary>
-	public required string Name { get; init; }
-
-	/// <summary>
-	/// The Value property, if set.
-	/// </summary>
-	public string? Value { get; init; }
-
-	/// <summary>
-	/// The Resource property, if set.
-	/// </summary>
-	public string? Resource { get; init; }
-}
-
-/// <summary>
-/// Assembly-level manifest attributes collected from all scanned assemblies.
-/// Aggregated across assemblies — used to generate top-level manifest elements
-/// like <![CDATA[<uses-permission>]]>, <![CDATA[<uses-feature>]]>, etc.
-/// </summary>
-public sealed class AssemblyManifestInfo
-{
-	public List<PermissionInfo> Permissions { get; } = [];
-	public List<PermissionGroupInfo> PermissionGroups { get; } = [];
-	public List<PermissionTreeInfo> PermissionTrees { get; } = [];
-	public List<UsesPermissionInfo> UsesPermissions { get; } = [];
-	public List<UsesFeatureInfo> UsesFeatures { get; } = [];
-	public List<UsesLibraryInfo> UsesLibraries { get; } = [];
-	public List<UsesConfigurationInfo> UsesConfigurations { get; } = [];
-	public List<MetaDataInfo> MetaData { get; } = [];
-	public List<PropertyInfo> Properties { get; } = [];
-
-	/// <summary>
-	/// Assembly-level [Application] attribute properties (merged from all assemblies).
-	/// Null if no assembly-level [Application] attribute was found.
-	/// </summary>
-	public Dictionary<string, object?>? ApplicationProperties { get; set; }
-}
-
-public sealed record PermissionInfo
-{
-	public required string Name { get; init; }
-	public IReadOnlyDictionary<string, object?> Properties { get; init; } = new Dictionary<string, object?> ();
-}
-
-public sealed record PermissionGroupInfo
-{
-	public required string Name { get; init; }
-	public IReadOnlyDictionary<string, object?> Properties { get; init; } = new Dictionary<string, object?> ();
-}
-
-public sealed record PermissionTreeInfo
-{
-	public required string Name { get; init; }
-	public IReadOnlyDictionary<string, object?> Properties { get; init; } = new Dictionary<string, object?> ();
-}
-
-public sealed record UsesPermissionInfo
-{
-	public required string Name { get; init; }
-	public int? MaxSdkVersion { get; init; }
-}
-
-public sealed record UsesFeatureInfo
-{
-	/// <summary>
-	/// Feature name (e.g., "android.hardware.camera"). Null for GL ES version features.
-	/// </summary>
-	public string? Name { get; init; }
-
-	/// <summary>
-	/// OpenGL ES version (e.g., 0x00020000 for 2.0). Zero for named features.
-	/// </summary>
-	public int GLESVersion { get; init; }
-
-	public bool Required { get; init; } = true;
-}
-
-public sealed record UsesLibraryInfo
-{
-	public required string Name { get; init; }
-	public bool Required { get; init; } = true;
-}
-
-public sealed record UsesConfigurationInfo
-{
-	public bool ReqFiveWayNav { get; init; }
-	public bool ReqHardKeyboard { get; init; }
-	public string? ReqKeyboardType { get; init; }
-	public string? ReqNavigation { get; init; }
-	public string? ReqTouchScreen { get; init; }
-}
-
-public sealed record PropertyInfo
-{
-	public required string Name { get; init; }
-	public string? Value { get; init; }
-	public string? Resource { get; init; }
-}
