+++ src/Xamarin.Android.Build.Tasks/Properties/Resources.resx
@@ -609,13 +609,6 @@ Either change the value in the AndroidManifest.xml to match the $(SupportedOSPla
   <data name="XA4213" xml:space="preserve">
     <value>The type '{0}' must provide a public default constructor</value>
   </data>
-  <data name="XA4212" xml:space="preserve">
-    <value>There can be only one type with an [Application] attribute; found: {0}</value>
-    <comment>{0} - Comma-separated list of types with [Application] attribute</comment>
-  </data>
-  <data name="XA4217" xml:space="preserve">
-    <value>Application cannot have both a type with an [Application] attribute and an [assembly:Application] attribute.</value>
-  </data>
   <data name="XA4214" xml:space="preserve">
     <value>The managed type `{0}` exists in multiple assemblies: {1}. Please refactor the managed type names in these assemblies so that they are not identical.</value>
     <comment>{0} - The managed type name
